Implement a number of Java.sun.misc.Unsafe.* methods.
Of particular interest is the addition of VMStaticInstance, used to
enable us to return a reference to a static instance, even though a
static instance isn't a heap object (and prevents the underlying
static instance from being GC'd while this reference is live).

This is required for us to support staticFieldBase/staticFieldOffset.