Delete the driver's HostInfo class. This abstraction just never really did anything. The two big pieces of functionality it tried to provide was to cache the ToolChain objects for each target, and to figure out the exact target based on the flag set coming in to an invocation. However, it had a lot of flaws even with those goals:  - Neither of these have anything to do with the host, or its info.  - The HostInfo class was setup as a full blown class *hierarchy* with    a separate implementation for each "host" OS. This required    dispatching just to create the objects in the first place.  - The hierarchy claimed to represent the host, when in fact it was    based on the target OS.  - Each leaf in the hierarchy was responsible for implementing the flag    processing and caching, resulting in a *lot* of copy-paste code and    quite a few bugs.  - The caching was consistently done based on architecture alone, even    though *any* aspect of the targeted triple might change the behavior    of the configured toolchain.  - Flag processing was already being done in the Driver proper,    separating the flag handling even more than it already is.
Instead of this, we can simply have the dispatch logic in the Driver
which previously created a HostInfo object create the ToolChain objects.
Adding caching in the Driver layer is a tiny amount of code. Finally,
pulling the flag processing into the Driver puts it where it belongs and
consolidates it in one location.

The result is that two functions, and maybe 100 lines of new code
replace over 10 classes and 800 lines of code. Woot.

This also paves the way to introduce more detailed ToolChain objects for
various OSes without threading through a new HostInfo type as well, and
the accompanying boiler plate. That, of course, was the yak I started to
shave that began this entire refactoring escapade. Wheee!